Update Create Collection Page
